selftest: iommu: fix hugepage size and command length type on riscv32 (#280)
This test sets the default hugepage size to 2MB but riscv32 uses 4MB hugepages by default.
This mismatch causes unmap failures during the test.

The condition to unmap IOAS is that the command length equals U64_MAX.
However, on riscv32, using size_t will never equal U64_MAX,
leading to the unmap operation being skipped.

Update the test to use the correct hugepage size and
ensure the command length is properly set using the correct type.

Reviewed-on: https://gitea.andestech.com/RD-SW/linux/pulls/280
Co-committed-by: Cynthia Huang <cynthia@andestech.com>

+++ tools/testing/selftests/iommu/iommufd_utils.h
@@ -279,7 +279,7 @@ static int _test_ioctl_ioas_map(int fd, unsigned int ioas_id, void *buffer,
 	})
 
 static int _test_ioctl_ioas_unmap(int fd, unsigned int ioas_id, uint64_t iova,
-				  size_t length, uint64_t *out_len)
+				  uint64_t length, uint64_t *out_len)
 {
 	struct iommu_ioas_unmap cmd = {
 		.size = sizeof(cmd),
